refactor(gauge): reworked label placement for gauge
If 'gauge_fullCircle: true':
The min label is positioned at the start angle within the inner circle.
The min label is positioned at the end angle within the inner circle and
is displayed as long as it does not overlap the min label.

If 'gauge_fullCircle: false':
The min label is positioned at the starting angle in the center of the arc.
The min label is positioned at the end angle in the center of
the arc and is displayed as long as it does not overlap the min label.